reorganize and consolidate
